HSEARCH-1592 Moving DefaultStringBridge to public API. Makes sense in the context of the bridges which already exists in this package
